+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-1.0.3 (unreleased)
+1.1.0 (2019-04-18)
 ------------------
 
 - Move stylesheets from head to body.